Backend/requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
django-mptt==0.14.0
django-mptt-admin==2.5.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
python-dotenv==1.0.0
drf_yasg==1.21.5
unidecode==1.3.6
django-cors-headers==4.0.0
